+++ src/axom/inlet/Container.cpp
@@ -213,10 +213,12 @@ Container& Container::addStructCollection(const std::string& name,
     std::string fullName = appendPrefix(m_name, name);
     fullName = removeAllInstances(fullName, detail::COLLECTION_GROUP_NAME + "/");
     detail::updateUnexpectedNames(fullName, m_unexpected_names);
-    if(m_reader.getIndices(fullName, indices))
+    const auto result = m_reader.getIndices(fullName, indices);
+    if(result == ReaderResult::Success)
     {
       container.addIndicesGroup(indices, description);
     }
+    markRetrievalStatus(*container.m_sidreGroup, result);
     markAsStructCollection(*container.m_sidreGroup);
   }
   return container;
@@ -387,10 +389,12 @@ axom::sidre::DataTypeId Container::addPrimitiveHelper<bool>(
   bool forArray,
   bool val)
 {
-  if(forArray || m_reader.getBool(lookupPath, val))
+  const auto result = m_reader.getBool(lookupPath, val);
+  if(forArray || result == ReaderResult::Success)
   {
     sidreGroup->createViewScalar("value", val ? int8(1) : int8(0));
   }
+  markRetrievalStatus(*sidreGroup, result);
   return axom::sidre::DataTypeId::INT8_ID;
 }
 
@@ -401,10 +405,12 @@ axom::sidre::DataTypeId Container::addPrimitiveHelper<int>(
   bool forArray,
   int val)
 {
-  if(forArray || m_reader.getInt(lookupPath, val))
+  const auto result = m_reader.getInt(lookupPath, val);
+  if(forArray || result == ReaderResult::Success)
   {
     sidreGroup->createViewScalar("value", val);
   }
+  markRetrievalStatus(*sidreGroup, result);
   return axom::sidre::DataTypeId::INT_ID;
 }
 
@@ -415,10 +421,12 @@ axom::sidre::DataTypeId Container::addPrimitiveHelper<double>(
   bool forArray,
   double val)
 {
-  if(forArray || m_reader.getDouble(lookupPath, val))
+  const auto result = m_reader.getDouble(lookupPath, val);
+  if(forArray || result == ReaderResult::Success)
   {
     sidreGroup->createViewScalar("value", val);
   }
+  markRetrievalStatus(*sidreGroup, result);
   return axom::sidre::DataTypeId::DOUBLE_ID;
 }
 
@@ -429,10 +437,12 @@ axom::sidre::DataTypeId Container::addPrimitiveHelper<std::string>(
   bool forArray,
   std::string val)
 {
-  if(forArray || m_reader.getString(lookupPath, val))
+  const auto result = m_reader.getString(lookupPath, val);
+  if(forArray || result == ReaderResult::Success)
   {
     sidreGroup->createViewString("value", val);
   }
+  markRetrievalStatus(*sidreGroup, result);
   return axom::sidre::DataTypeId::CHAR8_STR_ID;
 }
 
@@ -440,7 +450,7 @@ namespace detail
 {
 /*!
   *****************************************************************************
-  * \brief Adds the contents of an array to the table
+  * \brief Adds the contents of an array to the container
   * 
   * \return The keys that were added
   *****************************************************************************
@@ -460,7 +470,7 @@ std::vector<VariantKey> registerCollection(Container& container,
 
 /*!
   *****************************************************************************
-  * \brief Adds the contents of a dict to the table
+  * \brief Adds the contents of a dict to the container
   * 
   * \return The keys that were added
   *****************************************************************************
@@ -516,7 +526,8 @@ struct PrimitiveArrayHelper<Key, bool>
   {
     std::unordered_map<Key, bool> map;
     // Failure to retrieve a map is not necessarily an error
-    reader.getBoolMap(lookupPath, map);
+    const auto result = reader.getBoolMap(lookupPath, map);
+    markRetrievalStatus(*container.sidreGroup(), result);
     return registerCollection(container, map);
   }
 };
@@ -529,7 +540,8 @@ struct PrimitiveArrayHelper<Key, int>
                                      const std::string& lookupPath)
   {
     std::unordered_map<Key, int> map;
-    reader.getIntMap(lookupPath, map);
+    const auto result = reader.getIntMap(lookupPath, map);
+    markRetrievalStatus(*container.sidreGroup(), result);
     return registerCollection(container, map);
   }
 };
@@ -542,7 +554,8 @@ struct PrimitiveArrayHelper<Key, double>
                                      const std::string& lookupPath)
   {
     std::unordered_map<Key, double> map;
-    reader.getDoubleMap(lookupPath, map);
+    const auto result = reader.getDoubleMap(lookupPath, map);
+    markRetrievalStatus(*container.sidreGroup(), result);
     return registerCollection(container, map);
   }
 };
@@ -555,7 +568,8 @@ struct PrimitiveArrayHelper<Key, std::string>
                                      const std::string& lookupPath)
   {
     std::unordered_map<Key, std::string> map;
-    reader.getStringMap(lookupPath, map);
+    const auto result = reader.getStringMap(lookupPath, map);
+    markRetrievalStatus(*container.sidreGroup(), result);
     return registerCollection(container, map);
   }
 };
@@ -839,41 +853,48 @@ Container& Container::registerVerifier(std::function<bool(const Container&)> lam
 
 bool Container::verify() const
 {
-  bool verified = true;
+  // Whether the calling container has any "truthy" subcontainers, fields, or functions
+  // If the name is empty then we're the global (root) container, which we always
+  // consider to be defined
+  const bool this_container_defined = static_cast<bool>(*this) || m_name.empty();
+
   // If this container was required, make sure something was defined in it
-  verified &=
-    verifyRequired(*m_sidreGroup, static_cast<bool>(*this), "Container");
+  bool verified =
+    verifyRequired(*m_sidreGroup, this_container_defined, "Container");
+
   // Verify this Container if a lambda was configured
-  if(m_verifier && !m_verifier(*this))
+  if(this_container_defined && m_verifier && !m_verifier(*this))
   {
     verified = false;
     SLIC_WARNING(
       fmt::format("[Inlet] Container failed verification: {0}", m_name));
   }
-  // Verify the child Fields of this Container
-  for(const auto& field : m_fieldChildren)
-  {
-    if(!field.second->verify())
-    {
-      verified = false;
-    }
-  }
-  // Verify the child Containers of this Container
-  for(const auto& container : m_containerChildren)
-  {
-    if(!container.second->verify())
-    {
-      verified = false;
-    }
-  }
 
-  // Verify the child Functions of this Container
-  for(const auto& function : m_functionChildren)
+  // Checking the child objects is not needed if the container is empty
+  if(this_container_defined)
   {
-    if(!function.second->verify())
+    // Verify the child Fields of this Container
+    for(const auto& field : m_fieldChildren)
     {
-      verified = false;
+      verified = verified && field.second->verify();
     }
+    // Verify the child Containers of this Container
+    for(const auto& container : m_containerChildren)
+    {
+      verified = verified && container.second->verify();
+    }
+
+    // Verify the child Functions of this Container
+    for(const auto& function : m_functionChildren)
+    {
+      verified = verified && function.second->verify();
+    }
+  }
+  // If this has a collection group, it always needs to be verified, as annotations
+  // may have been applied to the collection group and not the calling group
+  else if(hasContainer(detail::COLLECTION_GROUP_NAME))
+  {
+    verified = verified && getContainer(detail::COLLECTION_GROUP_NAME).verify();
   }
 
   return verified;